pyproject.toml
[tool.poetry]
name = "tyuiuprofileservie"
version = "0.1.0"
description = ""
authors = ["AndreyKosov <andrey.kosov.05@inbox.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = {extras = ["all"], version = "^0.115.5"}
aiohttp = "^3.11.6"
sqlalchemy = "^2.0.36"
python-dotenv = "^1.0.1"
asyncpg = "^0.30.0"
alembic = "^1.14.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
